refactor: change helm-generic plugin to helm-installer
