--- docs/plugins/helm-generic.md
@@ -1,13 +1,0 @@
-# helm-generic Plugin
-
-// TODO
-
-## Usage
-
-The following content is an example of the "tool file".
-
-For more information on the main config, the tool file and the var file of DevStream, see [Core Concepts Overview](../core-concepts/core-concepts.md#1-config) and [DevStream Configuration](../core-concepts/config.md).
-
-``` yaml
---8<-- "helm-generic.yaml"
-```
