--- docs/plugins/helm-generic.zh.md
@@ -1,13 +1,0 @@
-# helm-generic 插件
-
-// TODO
-
-## 用例
-
-下面的配置文件展示的是"tool file"的内容。
-
-关于更多关于DevStream的主配置、tool file、var file的信息，请阅读[核心概念概览](../core-concepts/core-concepts.zh.md)和[DevStream配置](../core-concepts/config.zh.md).
-
-``` yaml
---8<-- "helm-generic.yaml"
-```
