+++ docs/plugins/helm-installer.md
@@ -1,0 +1,9 @@
+# helm-installer plugin
+
+TODO(daniel-hutao): I'll add the document after code refactoring finished.
+
+## Usage
+
+``` yaml
+--8<-- "helm-installer.yaml"
+```
