+++ docs/plugins/helm-installer.zh.md
@@ -1,0 +1,9 @@
+# helm-installer 插件
+
+TODO(daniel-hutao): I'll add the document after code refactoring finished.
+
+## 用例
+
+``` yaml
+--8<-- "helm-installer.yaml"
+```
