--- internal/pkg/plugin/helmgeneric/helmgeneric.go
@@ -1,1 +1,0 @@
-package helmgeneric
