+++ internal/pkg/plugin/helminstaller/update.go
@@ -1,4 +1,4 @@
-package helmgeneric
+package helminstaller
 
 import (
 	"github.com/devstream-io/devstream/internal/pkg/configmanager"
@@ -8,16 +8,15 @@ import (
 	"github.com/devstream-io/devstream/pkg/util/log"
 )
 
-// Create creates helmgeneric with provided options.
-func Create(options configmanager.RawOptions) (statemanager.ResourceStatus, error) {
+func Update(options configmanager.RawOptions) (statemanager.ResourceStatus, error) {
 	// Initialize Operator with Operations
 	operator := &plugininstaller.Operator{
 		PreExecuteOperations: plugininstaller.PreExecuteOperations{
 			helm.Validate,
 		},
-		ExecuteOperations:   helm.DefaultCreateOperations,
+		ExecuteOperations:   helm.DefaultUpdateOperations,
 		TerminateOperations: helm.DefaultTerminateOperations,
-		GetStatusOperation:  getEmptyStatus,
+		GetStatusOperation:  helm.GetAllResourcesStatus,
 	}
 
 	// Execute all Operations in Operator
