+++ internal/pkg/show/config/embed_gen.go
@@ -57,8 +57,8 @@ var (
 	//go:embed plugins/hashicorp-vault.yaml
 	HashicorpVaultDefaultConfig string
 
-	//go:embed plugins/helm-generic.yaml
-	HelmGenericDefaultConfig string
+	//go:embed plugins/helm-installer.yaml
+	HelmInstallerDefaultConfig string
 
 	//go:embed plugins/jenkins-pipeline.yaml
 	JenkinsPipelineDefaultConfig string
@@ -111,7 +111,7 @@ var pluginDefaultConfigs = map[string]string{
 	"harbor-docker":        HarborDockerDefaultConfig,
 	"harbor":               HarborDefaultConfig,
 	"hashicorp-vault":      HashicorpVaultDefaultConfig,
-	"helm-generic":         HelmGenericDefaultConfig,
+	"helm-installer":       HelmInstallerDefaultConfig,
 	"jenkins-pipeline":     JenkinsPipelineDefaultConfig,
 	"jenkins":              JenkinsDefaultConfig,
 	"jira-github-integ":    JiraGithubIntegDefaultConfig,
